[core] Add entry point for common compilation flow
