+++ package.json
@@ -20,15 +20,15 @@
     "@dev-sam/eslint-config-common": "0.21.0",
     "@dev-sam/jest-esbuild-transformer": "0.4.0",
     "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.0",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "esbuild": "^0.13.6",
+    "@types/node": "^16.11.1",
+    "@typescript-eslint/eslint-plugin": "^5.1.0",
+    "@typescript-eslint/parser": "^5.1.0",
+    "esbuild": "^0.13.8",
     "eslint": "^8.0.1",
     "eslint-import-resolver-node": "^0.3.6",
     "eslint-plugin-import": "^2.25.2",
     "husky": "^7.0.2",
-    "jest": "^27.2.5",
+    "jest": "^27.3.1",
     "lint-staged": "^11.2.3",
     "prettier": "^2.4.1",
     "source-map-support": "^0.5.20",
@@ -51,7 +51,6 @@
                 "**/samlang-core/checker",
                 "**/samlang-core/compiler",
                 "**/samlang-core/errors",
-                "**/samlang-core/interpreter/*",
                 "**/samlang-core/optimization",
                 "**/samlang-core/parser",
                 "**/samlang-core/printer",
