+++ samlang-cli/package.json
@@ -16,9 +16,9 @@
   "devDependencies": {
     "@dev-sam/samlang-core": "workspace:0.5.0",
     "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.0",
+    "@types/node": "^16.11.1",
     "@yarnpkg/esbuild-plugin-pnp": "^2.0.0-rc.1",
-    "esbuild": "^0.13.6",
+    "esbuild": "^0.13.8",
     "typescript": "^4.4.3",
     "vscode-languageserver": "^7.0.0"
   }
