--- samlang-cli/src/cli/index.ts
@@ -1,36 +1,0 @@
-import parseCLIArguments from './cli-parser';
-
-export interface CLIRunners {
-  format(needHelp: boolean): void;
-  typeCheck(needHelp: boolean): void;
-  compile(needHelp: boolean): void;
-  lsp(needHelp: boolean): void;
-  version(): void;
-  help(): void;
-}
-
-export default function cliMainRunner(
-  runners: CLIRunners,
-  commandLineArguments: readonly string[]
-): void {
-  const action = parseCLIArguments(commandLineArguments);
-  switch (action.type) {
-    case 'format':
-      runners.format(action.needHelp);
-      return;
-    case 'check':
-      runners.typeCheck(action.needHelp);
-      return;
-    case 'compile':
-      runners.compile(action.needHelp);
-      return;
-    case 'lsp':
-      runners.lsp(action.needHelp);
-      return;
-    case 'version':
-      runners.version();
-      return;
-    case 'help':
-      runners.help();
-  }
-}
