+++ samlang-cli/src/configuration.ts
@@ -1,8 +1,35 @@
 import { existsSync, readFileSync } from 'fs';
 import { dirname, join, resolve } from 'path';
 
-import parseSamlangProjectConfiguration from './configuration-parser';
-import type { SamlangProjectConfiguration } from './configuration-type';
+export type SamlangProjectConfiguration = {
+  readonly sourceDirectory: string;
+  readonly outputDirectory: string;
+  readonly entryPoints: readonly string[];
+};
+
+export function parseSamlangProjectConfiguration(
+  configurationString: string
+): SamlangProjectConfiguration | null {
+  try {
+    const json: unknown = JSON.parse(configurationString);
+    if (typeof json !== 'object' || json === null) return null;
+    const {
+      sourceDirectory = '.',
+      outputDirectory = 'out',
+      entryPoints = [],
+    } = json as Record<string, unknown>;
+    if (typeof sourceDirectory !== 'string' || typeof outputDirectory !== 'string') return null;
+    if (!Array.isArray(entryPoints)) return null;
+    const validatedEntryPoints: string[] = [];
+    entryPoints.forEach((entryPoint) => {
+      if (typeof entryPoint === 'string') validatedEntryPoints.push(entryPoint);
+    });
+    if (validatedEntryPoints.length !== entryPoints.length) return null;
+    return { sourceDirectory, outputDirectory, entryPoints: validatedEntryPoints };
+  } catch {
+    return null;
+  }
+}
 
 // Used for mock.
 type ConfigurationLoader = {
