--- samlang-cli/src/configuration/configuration-type.ts
@@ -1,5 +1,0 @@
-export type SamlangProjectConfiguration = {
-  readonly sourceDirectory: string;
-  readonly outputDirectory: string;
-  readonly entryPoints: readonly string[];
-};
