--- samlang-cli/src/configuration/index.ts
@@ -1,2 +1,0 @@
-export type { SamlangProjectConfiguration } from './configuration-type';
-export { default as loadSamlangProjectConfiguration } from './configuration-loader';
