--- samlang-core/interpreter/__tests__/panic-exception.test.ts
@@ -1,11 +1,0 @@
-import PanicException from '../panic-exception';
-
-describe('panic-exception', () => {
-  it('Throws Panic Exception', () => {
-    try {
-      throw new PanicException('panic!');
-    } catch (e: unknown) {
-      expect((e as { message: string }).message).toBe('panic!');
-    }
-  });
-});
