--- samlang-demo/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@dev-sam/samlang-demo",
-  "version": "0.4.4",
-  "license": "AGPLv3",
-  "files": [
-    "bin"
-  ],
-  "main": "./bin/index.js",
-  "types": "./bin/index.d.ts",
-  "scripts": {
-    "compile": "tsc",
-    "bundle": "node build.js"
-  },
-  "devDependencies": {
-    "@dev-sam/samlang-core": "workspace:0.5.0",
-    "@types/jest": "^27.0.2",
-    "@yarnpkg/esbuild-plugin-pnp": "^2.0.0-rc.1",
-    "esbuild": "^0.13.6",
-    "typescript": "^4.4.3"
-  }
-}
